Add files via upload
Soleilware's TDS software, Pallet detection system is released with online real-time detection and offline demo functions. Various types of pallets could be detected and the center and pose angle of multiple pallets could be given with +/-5mm and 0.25 degree in 2 meters with short-response time.  This program could be impelmented to a robot or self-driving forklift, with full automation. Please contect me at qsong@soleilwares.com for further information.